test: Add (double) cast for other printf_float macro instance
Need casts to avoid warning from -Wdouble-promotion

+++ test/printf_scanf.c
@@ -49,7 +49,7 @@
 # define _WANT_IO_C99_FORMATS
 # define _WANT_IO_LONG_LONG
 # define _WANT_IO_POS_ARGS
-# define printf_float(x) (x)
+# define printf_float(x) ((double) (x))
 #elif defined(TINY_STDIO)
 # ifdef PICOLIBC_INTEGER_PRINTF_SCANF
 #  ifndef _WANT_IO_POS_ARGS
